Remove some redundant logging and add an experimental GC callback helper.
Note we know relying on GC is bad design, but we might want to add some
additional asserts or logging on GC. Resource pools are sometimes hard to
make memory safe (e.g. C3P0 will leave threads running if we don't
properly close the pool, and it would be nice to be able to detect code
paths that might leak).